Crate rtnetlink

This crate provides methods to manipulate networking resources (links, addresses, arp tables, route tables) via the netlink protocol.

Re-exports

pub use netlink_packet_core as packet_core;

pub use netlink_packet_route as packet_route;

pub use netlink_proto as proto;

pub use netlink_sys as sys;

Modules

constants

Macros

try_nl

try_rtnl

Structs

AddressAddRequest

A request to create a new address. This is equivalent to the ip address add commands.

AddressDelRequest

AddressGetRequest

AddressHandle

AddressMessageBuilder

Helper struct for building AddressMessage.

Handle

LinkAddRequest

A request to create a new link. This is equivalent to the ip link add commands.

LinkBond

Represent bond interface. Example code on creating a bond interface

LinkBondPort

LinkBridge

Represent Linux Bridge interface. Example code on creating a linux bridge interface

LinkBridgePort

LinkBridgeVlan

LinkDelPropRequest

LinkDelRequest

LinkDummy

Represent dummy interface. Example code on creating a dummy interface

LinkGetRequest

LinkHandle

LinkMacSec

Represent MACsec interface. Example code on creating a MACsec interface

LinkMacVlan

Represent MAC VLAN interface. Example code on creating a MAC VLAN interface

LinkMacVtap

Represent MAC VTAP interface. Example code on creating a MAC VTAP interface

LinkMessageBuilder

Helper struct for building LinkMessage. The LinkMessageBuilder is designed for advanced user, wrapper structs/functions are created

LinkNetkit

Represent netkit virtual interface. Netkit devices are used for container networking and BPF programs.

LinkSetRequest

LinkUnspec

Generic interface without interface type Could be used to match interface by interface name or index. Example on attaching a interface to controller

LinkVeth

Represent virtual ethernet interface. Example code on creating a veth pair

LinkVlan

Represent VLAN interface. Example code on creating a VLAN interface

LinkVrf

Represent VRF interface. Example code on creating a VRF interface

LinkVxlan

Represent VxLAN interface. Example code on creating a VxLAN interface

LinkWireguard

Represent wireguard interface. Example code on creating a wireguard interface

LinkXfrm

Represent XFRM interface. Example code on creating a XFRM interface

NeighbourAddRequest

NeighbourDelRequest

NeighbourGetRequest

NeighbourHandle

NetworkNamespace

QDiscDelRequest

QDiscGetRequest

QDiscHandle

QDiscNewRequest

QosMapping

A quality-of-service mapping between the internal priority from to the external vlan priority to.

RouteAddRequest

A request to create a new route. This is equivalent to the ip route add commands.

RouteDelRequest

RouteGetRequest

RouteHandle

RouteMessageBuilder

RouteNextHopBuilder

RuleAddRequest

A request to create a new rule. This is equivalent to the ip rule add command.

RuleDelRequest

RuleGetRequest

RuleHandle

TrafficChainGetRequest

TrafficChainHandle

TrafficClassGetRequest

TrafficClassHandle

TrafficFilterGetRequest

TrafficFilterHandle

TrafficFilterNewRequest

Enums

Error

IpVersion

Internet Protocol (IP) version.

MulticastGroup

Constants

NETNS_PATH

NONE_FS

SELF_NS_PATH

Functions

from_socket

new_connection

new_connection_with_socket

new_multicast_connection

Equal to ip monitor command

new_multicast_connection_with_socket

Equal to ip monitor command