rtnetlink

Struct LinkMessageBuilder

pub struct LinkMessageBuilder<T> { /* private fields */ }

Helper struct for building LinkMessage. The LinkMessageBuilder is designed for advanced user, wrapper structs/functions are created

Implementations

impl LinkMessageBuilder<LinkBond>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for bond

pub fn append_info_data(self, info: InfoBond) -> Self

pub fn mode(self, mode: BondMode) -> Self

Adds the mode attribute to the bond This is equivalent to ip link add name NAME type bond mode MODE.

pub fn active_port(self, active_port: u32) -> Self

Adds the active_port attribute to the bond, where active_port is the ifindex of an interface attached to the bond. This is equivalent to ip link add name NAME type bond active_slave ACTIVE_PORT_NAME.

pub fn miimon(self, miimon: u32) -> Self

Adds the miimon attribute to the bond This is equivalent to ip link add name NAME type bond miimon MIIMON.

pub fn updelay(self, updelay: u32) -> Self

Adds the updelay attribute to the bond This is equivalent to ip link add name NAME type bond updelay UPDELAY.

pub fn downdelay(self, downdelay: u32) -> Self

Adds the downdelay attribute to the bond This is equivalent to ip link add name NAME type bond downdelay DOWNDELAY.

pub fn use_carrier(self, use_carrier: bool) -> Self

Adds the use_carrier attribute to the bond This is equivalent to ip link add name NAME type bond use_carrier USE_CARRIER.

pub fn arp_interval(self, arp_interval: u32) -> Self

Adds the arp_interval attribute to the bond This is equivalent to ip link add name NAME type bond arp_interval ARP_INTERVAL.

pub fn arp_validate(self, arp_validate: BondArpValidate) -> Self

Adds the arp_validate attribute to the bond This is equivalent to ip link add name NAME type bond arp_validate ARP_VALIDATE.

pub fn arp_all_targets(self, arp_all_targets: BondArpAllTargets) -> Self

Adds the arp_all_targets attribute to the bond This is equivalent to ip link add name NAME type bond arp_all_targets ARP_ALL_TARGETS

pub fn primary(self, primary: u32) -> Self

Adds the primary attribute to the bond, where primary is the ifindex of an interface. This is equivalent to ip link add name NAME type bond primary PRIMARY_NAME

pub fn primary_reselect(self, primary_reselect: BondPrimaryReselect) -> Self

Adds the primary_reselect attribute to the bond This is equivalent to ip link add name NAME type bond primary_reselect PRIMARY_RESELECT.

pub fn fail_over_mac(self, fail_over_mac: BondFailOverMac) -> Self

Adds the fail_over_mac attribute to the bond This is equivalent to ip link add name NAME type bond fail_over_mac FAIL_OVER_MAC.

pub fn xmit_hash_policy(self, xmit_hash_policy: BondXmitHashPolicy) -> Self

Adds the xmit_hash_policy attribute to the bond This is equivalent to ip link add name NAME type bond xmit_hash_policy XMIT_HASH_POLICY.

pub fn resend_igmp(self, resend_igmp: u32) -> Self

Adds the resend_igmp attribute to the bond This is equivalent to ip link add name NAME type bond resend_igmp RESEND_IGMP.

pub fn num_peer_notif(self, num_peer_notif: u8) -> Self

Adds the num_peer_notif attribute to the bond This is equivalent to ip link add name NAME type bond num_peer_notif NUM_PEER_NOTIF.

pub fn all_ports_active(self, all_ports_active: BondAllPortActive) -> Self

Adds the all_ports_active attribute to the bond This is equivalent to ip link add name NAME type bond all_slaves_active ALL_PORTS_ACTIVE.

pub fn min_links(self, min_links: u32) -> Self

Adds the min_links attribute to the bond This is equivalent to ip link add name NAME type bond min_links MIN_LINKS.

pub fn lp_interval(self, lp_interval: u32) -> Self

Adds the lp_interval attribute to the bond This is equivalent to ip link add name NAME type bond lp_interval LP_INTERVAL.

pub fn packets_per_port(self, packets_per_port: u32) -> Self

Adds the packets_per_port attribute to the bond This is equivalent to ip link add name NAME type bond packets_per_slave PACKETS_PER_PORT.

pub fn ad_lacp_rate(self, ad_lacp_rate: BondLacpRate) -> Self

Adds the ad_lacp_rate attribute to the bond This is equivalent to ip link add name NAME type bond ad_lacp_rate AD_LACP_RATE.

pub fn ad_select(self, ad_select: u8) -> Self

Adds the ad_select attribute to the bond This is equivalent to ip link add name NAME type bond ad_select AD_SELECT.

pub fn ad_actor_sys_prio(self, ad_actor_sys_prio: u16) -> Self

Adds the ad_actor_sys_prio attribute to the bond This is equivalent to ip link add name NAME type bond ad_actor_sys_prio AD_ACTOR_SYS_PRIO.

pub fn ad_user_port_key(self, ad_user_port_key: u16) -> Self

Adds the ad_user_port_key attribute to the bond This is equivalent to ip link add name NAME type bond ad_user_port_key AD_USER_PORT_KEY.

pub fn ad_actor_system(self, ad_actor_system: [u8; 6]) -> Self

Adds the ad_actor_system attribute to the bond This is equivalent to ip link add name NAME type bond ad_actor_system AD_ACTOR_SYSTEM.

pub fn tlb_dynamic_lb(self, tlb_dynamic_lb: bool) -> Self

Adds the tlb_dynamic_lb attribute to the bond This is equivalent to ip link add name NAME type bond tlb_dynamic_lb TLB_DYNAMIC_LB.

pub fn peer_notif_delay(self, peer_notif_delay: u32) -> Self

Adds the peer_notif_delay attribute to the bond This is equivalent to ip link add name NAME type bond peer_notif_delay PEER_NOTIF_DELAY.

pub fn ad_lacp_active(self, ad_lacp_active: bool) -> Self

Adds the ad_lacp_active attribute to the bond This is equivalent to ip link add name NAME type bond ad_lacp_active AD_LACP_ACTIVE.

pub fn missed_max(self, missed_max: u8) -> Self

Adds the missed_max attribute to the bond This is equivalent to ip link add name NAME type bond missed_max MISSED_MAX.

pub fn arp_ip_target(self, arp_ip_target: Vec<Ipv4Addr>) -> Self

Adds the arp_ip_target attribute to the bond This is equivalent to ip link add name NAME type bond arp_ip_target LIST.

pub fn ns_ip6_target(self, ns_ip6_target: Vec<Ipv6Addr>) -> Self

Adds the ns_ip6_target attribute to the bond This is equivalent to ip link add name NAME type bond ns_ip6_target LIST.

impl LinkMessageBuilder<LinkBondPort>

pub fn append_info_data(self, info: InfoBondPort) -> Self

Append arbitrary InfoBondPort

pub fn queue_id(self, queue_id: u16) -> Self

Adds the queue_id attribute to the bond port This is equivalent to ip link set name NAME type bond_slave queue_id QUEUE_ID.

pub fn prio(self, prio: i32) -> Self

Adds the prio attribute to the bond port This is equivalent to ip link set name NAME type bond_slave prio PRIO.

impl LinkMessageBuilder<LinkBridge>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for linux bridge

pub fn append_info_data(self, info: InfoBridge) -> Self

pub fn set_boolean_opt(self, opt: BridgeBooleanOptionFlags, value: bool) -> Self

pub fn ageing_time(self, value: u32) -> Self

pub fn group_fwd_mask(self, value: u16) -> Self

pub fn group_address(self, value: [u8; 6]) -> Self

pub fn forward_delay(self, value: u32) -> Self

pub fn hello_time(self, value: u32) -> Self

pub fn max_age(self, value: u32) -> Self

pub fn stp_state(self, value: BridgeStpState) -> Self

pub fn mst_enabled(self, value: bool) -> Self

pub fn priority(self, value: u16) -> Self

pub fn no_linklocal_learn(self, value: bool) -> Self

pub fn fdb_local_vlan_0(self, value: bool) -> Self

pub fn fdb_max_learned(self, value: u32) -> Self

pub fn vlan_filtering(self, value: bool) -> Self

pub fn vlan_protocol(self, value: VlanProtocol) -> Self

pub fn vlan_default_pvid(self, value: u16) -> Self

pub fn vlan_stats_enabled(self, value: bool) -> Self

pub fn vlan_stats_per_port(self, value: bool) -> Self

pub fn mcast_snooping(self, value: bool) -> Self

pub fn mcast_vlan_snooping(self, value: bool) -> Self

pub fn mcast_router(self, value: BridgeMulticastRouterType) -> Self

pub fn mcast_query_use_ifaddr(self, value: bool) -> Self

pub fn mcast_querier(self, value: bool) -> Self

pub fn mcast_hash_max(self, value: u32) -> Self

pub fn mcast_last_member_count(self, value: u32) -> Self

pub fn mcast_startup_query_count(self, value: u32) -> Self

pub fn mcast_last_member_interval(self, value: u64) -> Self

pub fn mcast_membership_interval(self, value: u64) -> Self

pub fn mcast_querier_interval(self, value: u64) -> Self

pub fn mcast_query_interval(self, value: u64) -> Self

pub fn mcast_query_response_interval(self, value: u64) -> Self

pub fn mcast_startup_query_interval(self, value: u64) -> Self

pub fn mcast_stats_enabled(self, value: bool) -> Self

pub fn mcast_igmp_version(self, value: u8) -> Self

pub fn mcast_mld_version(self, value: u8) -> Self

pub fn nf_call_iptables(self, value: bool) -> Self

pub fn nf_call_ip6tables(self, value: bool) -> Self

pub fn nf_call_arptables(self, value: bool) -> Self

pub fn mdb_offload_fail_notification(self, value: bool) -> Self

impl LinkMessageBuilder<LinkBridgePort>

pub fn append_info_data(self, info: InfoBridgePort) -> Self

Append arbitrary InfoBridgePort

pub fn fdb_flush(self) -> Self

This is equivalent to ip link set name NAME type bridge_slave fdb_flush.

pub fn state(self, v: BridgePortState) -> Self

This is equivalent to ip link set name NAME type bridge_slave state STATE.

pub fn priority(self, v: u16) -> Self

This is equivalent to ip link set name NAME type bridge_slave priority PRIO.

pub fn cost(self, v: u32) -> Self

This is equivalent to ip link set name NAME type bridge_slave cost COST.

pub fn guard(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave guard { on | off }.

pub fn hairpin(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave hairpin { on | off }.

pub fn root_block(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave root_block { on | off }.

pub fn learning(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave learning { on | off }.

pub fn flood(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave flood { on | off }.

pub fn proxy_arp(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave proxy_arp { on | off }.

pub fn proxy_arp_wifi(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave proxy_arp_wifi { on | off }.

pub fn mcast_router(self, v: BridgeMulticastRouterType) -> Self

This is equivalent to ip link set name NAME type bridge_slave mcast_router MULTICAST_ROUTER.

pub fn mcast_fast_leave(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave mcast_fast_leave { on | off }. and ip link set name NAME type bridge_slave fastleave { on | off }.

pub fn bcast_flood(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave bcast_flood { on | off }.

pub fn mcast_flood(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave mcast_flood { on | off }.

pub fn mcast_to_unicast(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave mcast_to_unicast { on | off }.

pub fn group_fwd_mask(self, v: u16) -> Self

This is equivalent to ip link set name NAME type bridge_slave group_fwd_mask MASK.

pub fn neigh_suppress(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave neigh_suppress { on | off }.

pub fn neigh_vlan_suppress(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave \ neigh_vlan_suppress { on | off }.

pub fn vlan_tunnel(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave vlan_tunnel { on | off }.

pub fn isolated(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave isolated { on | off }.

pub fn locked(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave locked { on | off }.

pub fn mab(self, v: bool) -> Self

This is equivalent to ip link set name NAME type bridge_slave mab { on | off }.

pub fn backup_port(self, iface_index: u32) -> Self

This is equivalent to ip link set name NAME type bridge_slave backup_port DEVICE but only accept interface index. Setting to 0 equal to Self::nobackup_port().

pub fn nobackup_port(self) -> Self

This is equivalent to ip link set name NAME type bridge_slave nobackup_port

pub fn backup_nhid(self, v: u32) -> Self

This is equivalent to ip link set name NAME type bridge_slave backup_nhid NHID

impl LinkMessageBuilder<LinkBridgeVlan>

pub fn append_af_spec(self, spec: AfSpecBridge) -> Self

Append arbitrary LinkAttribute::AfSpecBridge

pub fn vlan(self, vid: u16, flags: BridgeVlanInfoFlags) -> Self

pub fn vlan_range_start(self, vid: u16, flags: BridgeVlanInfoFlags) -> Self

Helper function by adding BridgeVlanInfoFlags::RangeBegin automatically to flags

pub fn vlan_range_end(self, vid: u16, flags: BridgeVlanInfoFlags) -> Self

Helper function by adding BridgeVlanInfoFlags::RangeEnd automatically to flags

pub fn bridge_self(self) -> Self

Change VLAN of bridge itself Equal to the self argument in bridge vlan add dev br0 vid 11 self

impl<T> LinkMessageBuilder<T>

pub fn new_with_info_kind(info_kind: InfoKind) -> Self

pub fn set_header(self, header: LinkHeader) -> Self

Set arbitrary LinkHeader

pub fn append_extra_attribute(self, link_attr: LinkAttribute) -> Self

Append arbitrary LinkAttribute

pub fn set_info_data(self, info_data: InfoData) -> Self

Set arbitrary InfoData

pub fn up(self) -> Self

Set the link up (equivalent to ip link set dev DEV up)

pub fn down(self) -> Self

Set the link down (equivalent to ip link set dev DEV down)

pub fn promiscuous(self, enable: bool) -> Self

Enable or disable promiscious mode of the link with the given index (equivalent to ip link set dev DEV promisc on/off)

pub fn arp(self, enable: bool) -> Self

Enable or disable the ARP protocol of the link with the given index (equivalent to ip link set dev DEV arp on/off)

pub fn name(self, name: String) -> Self

pub fn mtu(self, mtu: u32) -> Self

Set the mtu of the link with the given index (equivalent to ip link set DEV mtu MTU)

pub fn index(self, index: u32) -> Self

Kernel index number of interface, used for querying, modifying or deleting existing interface.

pub fn interface_family(self, family: AddressFamily) -> Self

pub fn address(self, address: Vec<u8>) -> Self

Define the hardware address of the link when creating it (equivalent to ip link add NAME address ADDRESS)

pub fn setns_by_pid(self, pid: u32) -> Self

Move this network device into the network namespace of the process with the given pid.

pub fn setns_by_fd(self, fd: RawFd) -> Self

Move this network device into the network namespace corresponding to the given file descriptor.

pub fn link(self, index: u32) -> Self

The physical device to act operate on. (e.g. the parent interface of VLAN/VxLAN)

pub fn controller(self, ctrl_index: u32) -> Self

Define controller interface index (similar to ip link set NAME master CONTROLLER_NAME)

pub fn nocontroller(self) -> Self

Detach the link from its controller. This is equivalent to ip link set LINK nomaster. To succeed, the link that is being detached must be UP.

pub fn set_port_kind(self, port_kind: InfoPortKind) -> Self

pub fn set_port_data(self, port_data: InfoPortData) -> Self

Include port settings. The [LinkBondPort] and [LinkBridgePort] are the helper

pub fn build(self) -> LinkMessage

impl LinkMessageBuilder<LinkUnspec>

pub fn new() -> Self

impl LinkMessageBuilder<LinkDummy>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for dummy interface type

impl LinkMessageBuilder<LinkMacVlan>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for MAC VLAN

pub fn append_info_data(self, info: InfoMacVlan) -> Self

pub fn mode(self, mode: MacVlanMode) -> Self

impl LinkMessageBuilder<LinkMacVtap>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for Mac VTAP interface

pub fn append_info_data(self, info: InfoMacVtap) -> Self

pub fn mode(self, mode: MacVtapMode) -> Self

impl LinkMessageBuilder<LinkMacSec>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for MACSEC

pub fn append_info_data(self, info: InfoMacSec) -> Self

pub fn sci(self, sci: u64) -> Self

pub fn port(self, port: u16) -> Self

pub fn icv_len(self, icv_len: u8) -> Self

pub fn cipher_suite(self, cipher_suite: MacSecCipherId) -> Self

pub fn window(self, window: u32) -> Self

pub fn encoding_sa(self, encoding_sa: u8) -> Self

pub fn encrypt(self, encrypt: bool) -> Self

pub fn protect(self, protect: bool) -> Self

pub fn inc_sci(self, inc_sci: bool) -> Self

pub fn es(self, es: bool) -> Self

pub fn scb(self, scb: bool) -> Self

pub fn replay_protect(self, replay_protect: bool) -> Self

pub fn validation(self, validation: MacSecValidate) -> Self

pub fn offload(self, offload: MacSecOffload) -> Self

pub fn other(self, other: DefaultNla) -> Self

impl LinkMessageBuilder<LinkNetkit>

pub fn new(name: &str, peer: &str, mode: NetkitMode) -> Self

Create LinkMessageBuilder for netkit

pub fn peer(self, peer: &str) -> Self

Set the peer interface name

pub fn mode(self, mode: NetkitMode) -> Self

Set the netkit mode (L2 or L3)

pub fn primary(self, primary: bool) -> Self

Set the primary interface flag

pub fn policy(self, policy: NetkitPolicy) -> Self

Set the policy for the primary interface

pub fn peer_policy(self, policy: NetkitPolicy) -> Self

Set the policy for the peer interface

pub fn scrub(self, scrub: NetkitScrub) -> Self

Set the scrub settings for the primary interface

pub fn peer_scrub(self, scrub: NetkitScrub) -> Self

Set the scrub settings for the peer interface

pub fn headroom(self, headroom: u16) -> Self

Set the desired headroom

pub fn tailroom(self, tailroom: u16) -> Self

Set the desired tailroom

impl LinkMessageBuilder<LinkVeth>

pub fn new(name: &str, peer: &str) -> Self

Create LinkMessageBuilder for VETH

pub fn peer(self, peer: &str) -> Self

impl LinkMessageBuilder<LinkVlan>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for VLAN

pub fn append_info_data(self, info: InfoVlan) -> Self

pub fn id(self, vlan_id: u16) -> Self

VLAN ID

pub fn protocol(self, protocol: VlanProtocol) -> Self

VLAN protocol

pub fn flags(self, flags: VlanFlags, mask: VlanFlags) -> Self

VLAN flags

pub fn qos<I, E>(self, ingress_qos: I, egress_qos: E) -> Self
where I: IntoIterator<Item = QosMapping>, E: IntoIterator<Item = QosMapping>,

ingress QoS and egress QoS

impl LinkMessageBuilder<LinkVrf>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for VRF

pub fn append_info_data(self, info: InfoVrf) -> Self

pub fn table_id(self, table_id: u32) -> Self

VRF table ID

impl LinkMessageBuilder<LinkVxlan>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for VLAN

pub fn append_info_data(self, info: InfoVxlan) -> Self

pub fn id(self, id: u32) -> Self

VNI

pub fn dev(self, index: u32) -> Self

This is equivalent to devPHYS_DEV for ip link vxlan. Adds the dev attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI dev LINK, dev LINK - specifies the physical device to use for tunnel endpoint communication. But instead of specifing a link name (LINK), we specify a link index. Please be aware the LinkMessageBuilder::link() will not works for VxLAN.

pub fn port(self, port: u16) -> Self

Adds the dstport attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI dstport PORT. dstport PORT - specifies the UDP destination port to communicate to the remote VXLAN tunnel endpoint.

pub fn group(self, addr: Ipv4Addr) -> Self

Adds the group attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI group IPADDR, group IPADDR - specifies the multicast IP address to join. This function takes an IPv4 address WARNING: only one between remote and group can be present.

pub fn group6(self, addr: Ipv6Addr) -> Self

Adds the group attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI group IPADDR, group IPADDR - specifies the multicast IP address to join. This function takes an IPv6 address WARNING: only one between remote and group can be present.

pub fn remote(self, addr: Ipv4Addr) -> Self

Adds the remote attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI remote IPADDR, remote IPADDR - specifies the unicast destination IP address to use in outgoing packets when the destination link layer address is not known in the VXLAN device forwarding database. This function takes an IPv4 address. WARNING: only one between remote and group can be present.

pub fn remote6(self, addr: Ipv6Addr) -> Self

Adds the remote attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI remote IPADDR, remote IPADDR - specifies the unicast destination IP address to use in outgoing packets when the destination link layer address is not known in the VXLAN device forwarding database. This function takes an IPv6 address. WARNING: only one between remote and group can be present.

pub fn local(self, addr: Ipv4Addr) -> Self

Adds the local attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI local IPADDR, local IPADDR - specifies the source IP address to use in outgoing packets. This function takes an IPv4 address.

pub fn local6(self, addr: Ipv6Addr) -> Self

Adds the local attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI local IPADDR, local IPADDR - specifies the source IP address to use in outgoing packets. This function takes an IPv6 address.

pub fn tos(self, tos: u8) -> Self

Adds the tos attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI tos TOS. tos TOS - specifies the TOS value to use in outgoing packets.

pub fn ttl(self, ttl: u8) -> Self

Adds the ttl attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI ttl TTL. ttl TTL - specifies the TTL value to use in outgoing packets.

pub fn label(self, label: u32) -> Self

Adds the flowlabel attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI flowlabel LABEL. flowlabel LABEL - specifies the flow label to use in outgoing packets.

pub fn learning(self, learning: bool) -> Self

Adds the learning attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI \[no\]learning. [no]learning - specifies if unknown source link layer addresses and IP addresses are entered into the VXLAN device forwarding database.

pub fn ageing(self, seconds: u32) -> Self

Adds the ageing attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI ageing SECONDS. ageing SECONDS - specifies the lifetime in seconds of FDB entries learnt by the kernel.

pub fn limit(self, limit: u32) -> Self

Adds the maxaddress attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI maxaddress LIMIT. maxaddress LIMIT - specifies the maximum number of FDB entries.

pub fn port_range(self, min: u16, max: u16) -> Self

Adds the srcport attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI srcport MIN MAX. srcport MIN MAX - specifies the range of port numbers to use as UDP source ports to communicate to the remote VXLAN tunnel endpoint.

pub fn proxy(self, proxy: bool) -> Self

Adds the proxy attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI [no]proxy. [no]proxy - specifies ARP proxy is turned on.

pub fn rsc(self, rsc: bool) -> Self

Adds the rsc attribute to the VXLAN This is equivalent to ip link add name NAME type vxlan id VNI [no]rsc. [no]rsc - specifies if route short circuit is turned on.

pub fn l2miss(self, l2miss: bool) -> Self

This is equivalent to ip link add name NAME type vxlan id VNI [no]l2miss. [no]l2miss - specifies if netlink LLADDR miss notifications are generated.

pub fn l3miss(self, l3miss: bool) -> Self

This is equivalent to ip link add name NAME type vxlan id VNI [no]l3miss. [no]l3miss - specifies if netlink IP ADDR miss notifications are generated.

pub fn collect_metadata(self, collect_metadata: bool) -> Self

pub fn udp_csum(self, udp_csum: bool) -> Self

This is equivalent to ip link add name NAME type vxlan id VNI [no]udp_csum. [no]udpcsum - specifies if UDP checksum is calculated for transmitted packets over IPv4.

impl LinkMessageBuilder<LinkWireguard>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for wireguard

impl LinkMessageBuilder<LinkXfrm>

pub fn new(name: &str) -> Self

Create LinkMessageBuilder for XFRM

pub fn append_info_data(self, info: InfoXfrm) -> Self

pub fn if_id(self, if_id: u32) -> Self

This is equivalent to the if_id IF_ID in command ip link add name NAME type xfrm if_id IF_ID.

pub fn dev(self, iface_index: u32) -> Self

This is equivalent to the dev PHYS_DEV in command ip link add name NAME type xfm dev PHYS_DEV, only take the interface index.