rtnetlink

Struct RouteMessageBuilder

pub struct RouteMessageBuilder<T = IpAddr> { /* private fields */ }

Implementations

impl<T> RouteMessageBuilder<T>

pub fn input_interface(self, index: u32) -> Self

Sets the input interface index.

pub fn output_interface(self, index: u32) -> Self

Sets the output interface index.

pub fn output_mpls(self, labels: Vec<MplsLabel>) -> Self

Sets the output MPLS encapsulation labels.

pub fn multipath(self, nexthops: Vec<RouteNextHop>) -> Self

Sets multiple nexthop entries for the route.

pub fn priority(self, priority: u32) -> Self

Sets the route priority (metric)

pub fn table_id(self, table: u32) -> Self

Sets the route table ID.

Default is main route table.

pub fn mark(self, mark: u32) -> Self

Sets mark value on route.

pub fn protocol(self, protocol: RouteProtocol) -> Self

Sets the route protocol.

Default is static route protocol.

pub fn scope(self, scope: RouteScope) -> Self

Sets the route scope.

Default is universe route scope.

pub fn kind(self, kind: RouteType) -> Self

Sets the route kind.

Default is unicast route kind.

pub fn onlink(self) -> Self

Marks the next hop as directly reachable (on-link).

Indicates that the next hop is reachable without passing through a connected subnet.

pub fn get_mut(&mut self) -> &mut RouteMessage

Return a mutable reference to the request message.

pub fn build(self) -> RouteMessage

impl RouteMessageBuilder<Ipv4Addr>

pub fn new() -> Self

Create default RouteMessage with header set to:

route: RouteHeader::RT_TABLE_MAIN

protocol: RouteProtocol::Static

scope: RouteScope::Universe

kind: RouteType::Unicast

address_family: [AddressFamily::Inet4]

For using this message in querying routes, these settings are ignored unless NETLINK_GET_STRICT_CHK been enabled.

pub fn source_prefix(self, addr: Ipv4Addr, prefix_length: u8) -> Self

Sets the source address prefix.

pub fn pref_source(self, addr: Ipv4Addr) -> Self

Sets the preferred source address.

pub fn destination_prefix(self, addr: Ipv4Addr, prefix_length: u8) -> Self

Sets the destination address prefix.

pub fn gateway(self, addr: Ipv4Addr) -> Self

Sets the gateway (via) address.

pub fn via(self, addr: Ipv6Addr) -> Self

Sets the IPv6 gateway (via) address.

impl RouteMessageBuilder<Ipv6Addr>

pub fn new() -> Self

Create default RouteMessage with header set to:

route: RouteHeader::RT_TABLE_MAIN

protocol: RouteProtocol::Static

scope: RouteScope::Universe

kind: RouteType::Unicast

address_family: AddressFamily::Inet6

For using this message in querying routes, these settings are ignored unless NETLINK_GET_STRICT_CHK been enabled.

pub fn source_prefix(self, addr: Ipv6Addr, prefix_length: u8) -> Self

Sets the source address prefix.

pub fn pref_source(self, addr: Ipv6Addr) -> Self

Sets the preferred source address.

pub fn destination_prefix(self, addr: Ipv6Addr, prefix_length: u8) -> Self

Sets the destination address prefix.

pub fn gateway(self, addr: Ipv6Addr) -> Self

Sets the gateway (via) address.

impl RouteMessageBuilder<IpAddr>

pub fn new() -> Self

pub fn source_prefix( self, addr: IpAddr, prefix_length: u8, ) -> Result<Self, InvalidRouteMessage>

Sets the source address prefix.

pub fn pref_source(self, addr: IpAddr) -> Result<Self, InvalidRouteMessage>

Sets the preferred source address.

pub fn destination_prefix( self, addr: IpAddr, prefix_length: u8, ) -> Result<Self, InvalidRouteMessage>

Sets the destination address prefix.

pub fn gateway(self, addr: IpAddr) -> Result<Self, InvalidRouteMessage>

Sets the gateway (via) address.

impl RouteMessageBuilder<MplsLabel>

pub fn new() -> Self

Create default RouteMessage with header set to:

route: RouteHeader::RT_TABLE_MAIN

protocol: RouteProtocol::Static

scope: RouteScope::Universe

kind: RouteType::Unicast

address_family: AddressFamily::Mpls

For using this message in querying routes, these settings are ignored unless NETLINK_GET_STRICT_CHK been enabled.

pub fn label(self, label: MplsLabel) -> Self

Sets the destination MPLS label.

pub fn via(self, addr: IpAddr) -> Self

Sets the gateway (via) address.