rtnetlink

Struct RouteNextHopBuilder

pub struct RouteNextHopBuilder { /* private fields */ }

Implementations

impl RouteNextHopBuilder

pub fn new(address_family: AddressFamily) -> Self

Create default RouteNexthop for a route with the given address family.

pub fn new_ipv4() -> Self

Create IPv4 RouteNexthop

pub fn new_ipv6() -> Self

Create IPv6 RouteNexthop

pub fn interface(self, index: u32) -> Self

Sets the nexthop interface index.

pub fn via(self, addr: IpAddr) -> Result<Self, InvalidRouteMessage>

Sets the nexthop (via) address.

pub fn onlink(self) -> Self

Marks the nexthop as directly reachable (on-link).

Indicates that the nexthop is reachable without passing through a connected subnet.

pub fn mpls(self, labels: Vec<MplsLabel>) -> Self

Sets the nexthop MPLS encapsulation labels.

pub fn weight(self, weight: u8) -> Self

Set the nexthop weight

Equal to weight property in ip route, but please be advised the number shown in ip route command has plus 1. Meaning kernel has weight 0, but ip route shows as weight 1. This function is using kernel number from range of 0 to 255.

pub fn flags(self, flags: RouteNextHopFlags) -> Self

Set flags for next hop

pub fn build(self) -> RouteNextHop