Crate rtnetlink

This crate provides methods to manipulate networking resources (links, addresses, arp tables, route tables) via the netlink protocol.

Re-exports

pub use netlink_packet_route as packet;

Modules

constants

proto

sys

Macros

try_nl

try_rtnl

Structs

AddressAddRequest

A request to create a new address. This is equivalent to the ip address add commands.

AddressDelRequest

AddressGetRequest

AddressHandle

Handle

LinkAddRequest

A request to create a new link. This is equivalent to the ip link add commands.

LinkDelPropRequest

LinkDelRequest

LinkGetRequest

LinkHandle

LinkNewPropRequest

LinkSetRequest

NeighbourAddRequest

NeighbourDelRequest

NeighbourGetRequest

NeighbourHandle

NetworkNamespace

QDiscDelRequest

QDiscGetRequest

QDiscHandle

QDiscNewRequest

RouteAddRequest

A request to create a new route. This is equivalent to the ip route add commands.

RouteDelRequest

RouteGetRequest

RouteHandle

RuleAddRequest

A request to create a new rule. This is equivalent to the ip rule add command.

RuleDelRequest

RuleGetRequest

RuleHandle

TrafficChainGetRequest

TrafficChainHandle

TrafficClassGetRequest

TrafficClassHandle

TrafficFilterGetRequest

TrafficFilterHandle

TrafficFilterNewRequest

VxlanAddRequest

A request to create a new vxlan link. This is equivalent to ip link add NAME vxlan id ID ... commands. It provides methods to customize the creation of the vxlan interface It provides almost all parameters that are listed by man ip link.

Enums

Error

IpVersion

Internet Protocol (IP) version.

Constants

NETNS_PATH

NONE_FS

SELF_NS_PATH

Functions

new_connection

new_connection_with_socket