Module rtnetlink::sys::protocols

Constants

NETLINK_AUDIT

Auditing.

NETLINK_CONNECTOR

Kernel connector. See Documentation/connector/* in the Linux kernel source tree for further information.

NETLINK_CRYPTO

Netlink interface to request information about ciphers registered with the kernel crypto API as well as allow configuration of the kernel crypto API.

NETLINK_DNRTMSG

DECnet routing messages.

NETLINK_ECRYPTFS

NETLINK_FIB_LOOKUP

Access to FIB lookup from user space.

NETLINK_FIREWALL

Transport IPv4 packets from netfilter to user space. Used by ip_queue kernel module. After a long period of being declared obsolete (in favor of the more advanced nfnetlink_queue feature), it was removed in Linux 3.5.

NETLINK_GENERIC

Generic netlink family for simplified netlink usage.

NETLINK_IP6_FW

Transport IPv6 packets from netfilter to user space. Used by ip6_queue kernel module.

NETLINK_ISCSI

Open-iSCSI.

NETLINK_KOBJECT_UEVENT

Kernel messages to user space.

NETLINK_NETFILTER

Netfilter subsystem.

NETLINK_NFLOG

Netfilter/iptables ULOG.

NETLINK_RDMA

Infiniband RDMA.

NETLINK_ROUTE

Receives routing and link updates and may be used to modify the routing tables (both IPv4 and IPv6), IP addresses, link parameters, neighbor setups, queueing disciplines, traffic classes and packet classifiers (see rtnetlink(7)).

NETLINK_SCSITRANSPORT

SCSI transpots

NETLINK_SELINUX

SELinux event notifications.

NETLINK_SOCK_DIAG

Query information about sockets of various protocol families from the kernel (see sock_diag(7)).

NETLINK_UNUSED

NETLINK_USERSOCK

Reserved for user-mode socket protocols.

NETLINK_XFRM

IPsec.